--- backend/arduino/main.ino
@@ -1,14 +1,0 @@
-int led = 13;
-
-void setup()
-{
-  pinMode(led, OUTPUT);
-}
-
-void loop()
-{
-  digitalWrite(led, HIGH);
-  delay(1000);
-  digitalWrite(led, LOW);
-  delay(1000);
-}
