+++ backend/package.json
@@ -10,6 +10,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "johnny-five": "^1.4.0"
+    "johnny-five": "^1.4.0",
+    "pi-camera": "^1.3.0"
   }
 }
